Wait for one block before updating staked token balances
Staked token balances are counted based on token balance at the
beginning of the block. To get the staked token balance after we have
staked or unstaked tokens we need to wait for one block.

This has implications for determining if a user is a member. We now
properly handle the case where a user stakes some tokens and then
attempts to vote on a proposal as we will display a loading indicator
until the we have learned the most up-to-date staking balance.