Stop namespacing favorited contracts based on the code ID
This was a relic of the past when we were making backwards
incompatible changes on the testnet. Whereas previously we would
namespace based on the code ID, we now namespace based on a series of
backwards compatible ones. Should we make a non-backwards compatible
change we can update the `NEXT_PUBLIC_FAVORITES_NAMESPACE` to that
code id.