Fix issues with default theming
Previously we used a library called theme-change which via a strange
attribute on our theme button would change the theme of our webpage
when the button was pressed. It would then persist that information
into local storage.

Trouble is that becasue we use NextJs many things are rendered on
server side where a default theme is set on the html. If the theme
stored in local storage didn't match the theme served as a static file
some elements would render strangely.

We resolve this by getting rid of said library and just doing what it
did in a reasonably sane way that is compatible with nextJS.

This change also makes dark mode the default.