Include stargate and staking extensions in cosmos msg schema.
Generated by enabling the staking and stargate extensions in the
cosmwasm_std package and then running `cargo schema` to generate the
files.

Note that this makes a slight modification to the generated schema in
that it removes the usage of the uint64 format. JSON Schema does not
actually support such a type / format. It is unclear to me why this
appears in the generated schema with those extensions while it does
not without them.

For example, given the generated schema:

```json
"proposal_id": {
  "type": "integer",
  "format": "uint64",
  "minimum": 0.0
}
```

After the manual modification it will look like:

```json
"proposal_id": {
  "type": "integer",
  "minimum": 0.0
}
```

This matches the way the u64 types are turned into a schema without
those extensions enabled.