Disable check for staked balance at proposal creation time.
Currently there is an issue with querying past chain state from before
the chain halted. This makes voting on proposals created before the
chain halt not possible as that check requires those queries to work.

This removes that check for now. Next week, once those proposals have
cleared we can re enable it.