Properly handle non-native decimals in Multisig spend template.
Previously we relied on state from the DAO config which is not present
when a Multisig is being used. This modifies the spend template to
store an object in the form `{ address: string, decimals: number }`
when a selection is made. The decimals there are then used when
computing amounts.

This seems a little hacky, but we do in fact need to do it because we
can not fetch decimals information outside of components where the
transformation from template data to a cosmos message happens.