Deploy latest changes to main. (#529)
* Properly meter cw20 balances queries to stop out of gas issues.

* Update UI to use self hosted hero font. (#511)

* Properly handle non-native decimals in Multisig spend template.

Previously we relied on state from the DAO config which is not present
when a Multisig is being used. This modifies the spend template to
store an object in the form `{ address: string, decimals: number }`
when a selection is made. The decimals there are then used when
computing amounts.

This seems a little hacky, but we do in fact need to do it because we
can not fetch decimals information outside of components where the
transformation from template data to a cosmos message happens.

* Fix rendering of unknown IBC assets.

* latest ibc assets from junoswap

after some jq field capitalization matches

* Line wrap in markdown preview to improve mobile view.

* Remove beta notice popup that gets in the way of things.

Co-authored-by: bartolomeə <barton@bmorphism.xyz>